+++ cmake/modules/DumuxMacros.cmake
@@ -1,8 +1,10 @@
 # additional macros
+include(AddGstatFileLinks)
 include(AddInputFileLinks)
 include(DumuxDoxygen)
 include(DumuxTestMacros)
 
+find_package(Gstat)
 find_package(Gnuplot)
 set(HAVE_GNUPLOT ${GNUPLOT_FOUND})